requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0

# Reddit
praw>=7.7.1

# Google Gemini
google-genai>=1.0.0

# Telegram
python-telegram-bot>=20.7

# Image / video
Pillow>=10.2.0

# Google TTS + YouTube upload
google-cloud-texttospeech>=2.16.0
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
responses>=0.25.0
